Refactors Docker configuration for easier management
Simplifies and centralizes the Docker configuration by mounting individual configuration files as volumes instead of copying them during the build process. This allows for live reloading of configuration changes without rebuilding the Docker image.

Additionally, this commit reduces the number of installed PHP extensions in the Dockerfile to include only core extensions. This change, coupled with the configuration volume mounts, makes it easier to manage and customize PHP extensions.

Adds a `reload_services` function to the `ium` script to restart the docker compose services and reload the configurations, so that the changes are applied.